backend/Apis/requirements.txt
pipe==2.0
web3==5.31.3
sentry-sdk[fastapi]==1.30.0
pydantic==1.10.2
fastapi
uvicorn
tortoise-orm[asyncpg]==0.19.3
redis
pytz
cache
web3_multicall
fastapi_limiter
fastapi-cache2
requests[security]
retry_requests
opentelemetry-distro
opentelemetry-exporter-otlp
slowapi
retry